refactor: remove dead guardrails wrapper, cut redundancies, fix registry trim
Quality pass focused on removing dead/duplicated code and one routing bug,
while keeping behavior and test coverage intact.

fix(registry): provider types were stored verbatim at registration while
every reader except ProviderByType trimmed on lookup, so a configured type
with surrounding whitespace (YAML/env) made type-based routing silently
return nil. Normalize the type at registration; add a regression test.

refactor(guardrails): delete the dead GuardedProvider wrapper (provider.go,
518 lines) and the unused RequestPatcher/BatchPreparer in executor.go. The
live request path applies guardrails through WorkflowRequestPatcher /
WorkflowBatchPreparer (wired in app.go); the wrapper was only reachable from
its own tests. Live clone helpers moved to clone.go. Coverage of the shared
rewrite functions (processGuardedChat/Responses/BatchRequest) is preserved
via a trimmed test-only harness; the 6 pure-delegation tests were dropped and
the server tests retargeted onto the live Workflow* patchers.

refactor: assorted redundancy cleanups -
- providers/router.go: collapse 3 near-identical Native*ProviderTypes loops
  into one providerTypesSupporting helper.
- core/errors.go: add NewEmptyProviderResponseError, replacing 8 duplicated
  constructions in native_response_service.go and inference_execute.go.
- replace 4 hand-rolled map[string]any clone helpers with stdlib maps.Clone.
- anthropic: drop an unreachable len(tokens)==0 branch.

Net ~-469 lines. go build/vet/test and gofmt all clean.
