+++ internal/guardrails/executor.go
@@ -8,58 +8,6 @@ import (
 	"gomodel/internal/core"
 )
 
-// RequestPatcher applies guardrails to translated requests without owning
-// provider execution.
-type RequestPatcher struct {
-	pipeline *Pipeline
-}
-
-// NewRequestPatcher creates an explicit translated-request patcher.
-func NewRequestPatcher(pipeline *Pipeline) *RequestPatcher {
-	return &RequestPatcher{pipeline: pipeline}
-}
-
-// PatchChatRequest applies guardrails to a translated chat request.
-func (p *RequestPatcher) PatchChatRequest(ctx context.Context, req *core.ChatRequest) (*core.ChatRequest, error) {
-	return processGuardedChat(ctx, p.pipeline, req)
-}
-
-// PatchResponsesRequest applies guardrails to a translated responses request.
-func (p *RequestPatcher) PatchResponsesRequest(ctx context.Context, req *core.ResponsesRequest) (*core.ResponsesRequest, error) {
-	return processGuardedResponses(ctx, p.pipeline, req)
-}
-
-// BatchPreparer applies guardrails to native batch subrequests before provider
-// submission.
-type BatchPreparer struct {
-	provider core.RoutableProvider
-	pipeline *Pipeline
-}
-
-// NewBatchPreparer creates an explicit native-batch preparer.
-func NewBatchPreparer(provider core.RoutableProvider, pipeline *Pipeline) *BatchPreparer {
-	return &BatchPreparer{
-		provider: provider,
-		pipeline: pipeline,
-	}
-}
-
-// PrepareBatchRequest applies guardrails to batch subrequests without
-// submitting the batch to the wrapped provider.
-func (p *BatchPreparer) PrepareBatchRequest(ctx context.Context, providerType string, req *core.BatchRequest) (*core.BatchRewriteResult, error) {
-	return processGuardedBatchRequest(ctx, providerType, req, p.pipeline, p.batchFileTransport())
-}
-
-func (p *BatchPreparer) batchFileTransport() core.BatchFileTransport {
-	if p == nil || p.provider == nil {
-		return nil
-	}
-	if files, ok := p.provider.(core.NativeFileRoutableProvider); ok {
-		return files
-	}
-	return nil
-}
-
 func processGuardedBatchRequest(
 	ctx context.Context,
 	providerType string,
